fix: prevent Any narrowing after subscript assignment
