+++ pyrefly/lib/alt/solve.rs
@@ -5261,12 +5261,11 @@ impl<'a, Ans: LookupAnswer> AnswersSolver<'a, Ans> {
 
     /// Whether post-assignment narrowing of `arm[k]` to the assigned value is
     /// sound for this union arm. Defers to the per-class cached
-    /// `KeyClassSubscriptSymmetry` answer for `ClassType` arms. Explicit
-    /// `Any` is never narrowed, while other types preserve their existing
-    /// behavior.
+    /// `KeyClassSubscriptSymmetry` answer for `ClassType` arms. `Any` is never
+    /// narrowed, while other types preserve their existing behavior.
     fn subscript_assign_arm_allows_narrowing(&self, arm: &Type) -> bool {
         match arm {
-            Type::Any(AnyStyle::Explicit) => false,
+            Type::Any(_) => false,
             Type::ClassType(cls) => *self.get_subscript_symmetry_for_class(cls.class_object()),
             _ => true,
         }
