+++ pyrefly/lib/test/assign.rs
@@ -44,43 +44,6 @@ def f(x: int) -> None:
 );
 
 testcase!(
-    test_subscript_assign_does_not_narrow_any,
-    r#"
-from typing import Any, assert_type
-
-values: list[float] = []
-frame: Any = object()
-frame["start"] = values
-assert_type(frame["start"], Any)
-frame["end"] = frame["start"].shift(-1)
-"#,
-);
-
-testcase!(
-    test_subscript_assign_implicit_any_still_narrows,
-    r#"
-from typing import assert_type
-
-def f(frame):
-    values: list[float] = []
-    frame["start"] = values
-    assert_type(frame["start"], list[float])
-"#,
-);
-
-testcase!(
-    test_subscript_assign_error_any_still_narrows,
-    r#"
-from typing import assert_type
-
-def f(frame: MissingType) -> None:  # E: Could not find name `MissingType`
-    values: list[float] = []
-    frame["start"] = values
-    assert_type(frame["start"], list[float])
-"#,
-);
-
-testcase!(
     test_error_assign,
     r#"
 x: str = 1  # E: `Literal[1]` is not assignable to `str`
